Merge branch 'main' of https://github.com/wkmeyer-lab/RunRERBinaryMT

# Conflicts:
#	Src/Loc/RunPermulationsManual.R

+++ Src/Loc/MakeTreeBinaryManualAnnotations.R
@@ -2,6 +2,7 @@
 library(RERconverge) #load RERconverge package
 library(RERconverge)
 library(tools)
+library(devtools)
 source(file = "Src/Reu/cmdArgImport.R")
 
 # ---- USAGE,README ----
@@ -23,7 +24,7 @@ mainTreesLocation = "/share/ceph/wym219group/shared/projects/MammalDiet/Zoonomia
 
 #local computer debug version:
 # mainTreesLocation = "data/RemadeTreesAllZoonomiaSpecies.rds"
-#args = c("m=data/RemadeTreesAllZoonomiaSpecies.rds", "r=allInsectivory", "a=Ins_v_herbs", "t=bi", "c=all", "w=F")
+#args = c("m=data/RemadeTreesAllZoonomiaSpecies.rds", "r=carnvHerbs", "a=CarnFish_Herbs", "t=bi", "c=all", "w=F")
 
 #Other defaults if not specified
 annotCollumn = "ERRORDEFAULT"
@@ -121,16 +122,23 @@ if(!dir.exists("Results")){
 
 manualAnnots = read.csv("Data/manualAnnotationsSheet.csv")
 
-# --- subset the manual annots to only those with data in the collumn
-relevantSpecies = manualAnnots[ manualAnnots[[annotCollumn]] %in% c(0,1),]
-
-relevantSpeciesNames = relevantSpecies$FaName
-
-# -- output a species filter for this fileprefix 
-
+# ---- if a pre-defined species filter exists, use that. If not, create one. 
 speciesFilterFilename = paste(outputFolderName, filePrefix, "SpeciesFilter.rds",sep="")
-saveRDS(relevantSpeciesNames, file = speciesFilterFilename)
 
+
+if(!file.exists(speciesFilterFilename)){
+  # --- subset the manual annots to only those with data in the collumn
+  relevantSpecies = manualAnnots[ manualAnnots[[annotCollumn]] %in% c(0,1),]
+  
+  relevantSpeciesNames = relevantSpecies$FaName
+  
+  # -- output a species filter for this fileprefix 
+  
+  saveRDS(relevantSpeciesNames, file = speciesFilterFilename)
+}else{
+  relevantSpecieslist = readRDS(speciesFilterFilename)
+  relevantSpecies = manualAnnots[ manualAnnots[["FaName"]] %in% relevantSpecieslist,]
+}
 # -- Setup foreground Species -- 
 
 foregroundSpeciesAnnot = relevantSpecies[ relevantSpecies[[annotCollumn]] %in% 1,]
